Moved include tests into syntax_tests/test_include.py.
